separate prioritize struct
